replication(coordinator): refactor push and adjust timeout
