+++ usecases/replica/coordinator.go
@@ -153,112 +153,6 @@ func (c *coordinator[T]) broadcast(ctx context.Context,
 	return resChan
 }
 
-// commitAll tells replicas to commit pending updates related to a specific request
-// (second phase of a two-phase commit)
-func (c *coordinator[T]) commitAll(ctx context.Context,
-	broadcastCh <-chan _Result[string],
-	op commitOp[T],
-	callback func(successful int),
-) <-chan _Result[T] {
-	replyCh := make(chan _Result[T], cap(broadcastCh))
-	f := func() { // tells active replicas to commit
-		// tells active replicas to commit
-
-		var successful atomic.Int32
-
-		defer func() {
-			if callback != nil {
-				callback(int(successful.Load()))
-			}
-		}()
-
-		wg := sync.WaitGroup{}
-
-		for res := range broadcastCh {
-			if res.Err != nil {
-				replyCh <- _Result[T]{Err: res.Err}
-				continue
-			}
-			replica := res.Value
-			wg.Add(1)
-			g := func() {
-				defer wg.Done()
-				resp, err := op(ctx, replica, c.TxID)
-				if err == nil {
-					successful.Add(1)
-				}
-				replyCh <- _Result[T]{resp, err}
-			}
-			enterrors.GoWrapper(g, c.log)
-		}
-
-		wg.Wait()
-		close(replyCh)
-	}
-	enterrors.GoWrapper(f, c.log)
-
-	return replyCh
-}
-
-// Push pushes updates to all replicas of a specific shard
-func (c *coordinator[T]) Push(ctx context.Context,
-	cl types.ConsistencyLevel,
-	ask readyOp,
-	com commitOp[T],
-) (<-chan _Result[T], int, error) {
-	options := c.Router.BuildRoutingPlanOptions(c.Shard, c.Shard, cl, "")
-	writeRoutingPlan, err := c.Router.BuildWriteRoutingPlan(options)
-	if err != nil {
-		return nil, 0, fmt.Errorf("%w : class %q shard %q", err, c.Class, c.Shard)
-	}
-
-	level := writeRoutingPlan.IntConsistencyLevel
-
-	//nolint:govet // we expressely don't want to cancel that context as the timeout will take care of it
-	ctxWithTimeout, _ := context.WithTimeout(context.Background(), 20*time.Second)
-	c.log.WithFields(logrus.Fields{
-		"action":   "coordinator_push",
-		"duration": 20 * time.Second,
-		"level":    level,
-	}).Debug("context.WithTimeout")
-
-	// create callback for metrics
-	// the use of an immediately invoked function expression (IIFE) captures the start time
-	// and returns the actual callback function.
-	// The returned function is then called by commitAll once it knows how many
-	// replicas have successfully committed
-	callback := func() func(successful int) {
-		start := time.Now()
-
-		return func(successful int) {
-			numReplicas := len(writeRoutingPlan.Replicas())
-
-			if numReplicas == successful {
-				c.metrics.IncWritesSucceedAll()
-			} else if successful > 0 {
-				c.metrics.IncWritesSucceedSome()
-			} else {
-				c.metrics.IncWritesFailed()
-			}
-
-			c.metrics.ObserveWriteDuration(time.Since(start))
-		}
-	}()
-
-	nodeCh := c.broadcast(ctxWithTimeout, writeRoutingPlan.HostAddresses(), ask, level)
-	commitCh := c.commitAll(context.Background(), nodeCh, com, callback)
-
-	// if there are additional hosts, we do a "best effort" write to them
-	// where we don't wait for a response because they are not part of the
-	// replicas used to reach level consistency
-	if len(writeRoutingPlan.AdditionalHostAddresses()) > 0 {
-		additionalHostsBroadcast := c.broadcast(ctxWithTimeout, writeRoutingPlan.AdditionalHostAddresses(), ask, len(writeRoutingPlan.AdditionalHostAddresses()))
-		c.commitAll(context.Background(), additionalHostsBroadcast, com, nil)
-	}
-
-	return commitCh, level, nil
-}
-
 // Pull data from replica depending on consistency level, trying to reach level successful calls
 // to op, while cycling through replicas for the coordinator's shard.
 //
